같은 서비스를 양쪽이 각자 구현해 34개 파일이 충돌했다. 파일 단위로 풀면 두 설계가
섞이므로 서비스 단위로 어느 쪽을 쓸지 정하고 나머지를 옮겨 붙였다.

material-service — 서홍님 구현 채택
- ApiException 기반 예외 처리, UserClient, 서비스별 DB(material_db)
- main 쪽에서 가져온 것: 데모 원료 시드 10건(DataSeeder),
  내부 API 키 보호(InternalApiInterceptor)
- unitPrice 가 BigDecimal 이라 DataSeeder 를 맞춤
- deductCapacity 가 Math.max(0, ...) 로 부족분을 조용히 0 으로 깎고 있었다.
  동시 주문이 겹치면 여유분보다 많이 나가므로 부족하면 거절하도록 바꾼다
- 보상 경로용 capacity-restore 엔드포인트 추가

order-service — main 구현 유지
- 서홍님 버전에는 재고 기능이 없다. 브랜치를 딴 시점이 PR #1 머지 전이라
  통째로 바꾸면 재환님이 올린 재고·부족 감지 스케줄러가 사라진다
- 유지: 재고(PR #1), 결제 완료 이벤트 수신, 내부 API 키 보호
- 흡수: 서비스별 DB(order_db)
- material-service 호출을 서홍님 규약(capacity-deduct / capacity-restore)에 맞춤

payment-service — main 구현 유지 (조달 결제 · Kafka)

공통
- 게이트웨이: 서홍님 방식(gateway/routes.yml + SPRING_CONFIG_ADDITIONAL_LOCATION)으로
  일원화하고 api-gateway-config/ 를 제거한다.
  PR #1 의 /api/inventories/** 라우트가 빠지지 않도록 옮겨 담았다
- init-db/00_databases.sql: 서비스별 스키마 생성
- course-service 제거 (강의 템플릿 잔재. 참조하는 서비스 없음)
- .gitignore: 양쪽 규칙을 모두 유지

프론트엔드는 수정하지 않았다. 서홍님 DTO 가 기존 화면과 그대로 호환된다.

+++ msa-MyService/material-service/src/main/java/com/lecture/material/config/DataSeeder.java
@@ -9,6 +9,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -100,7 +101,7 @@ public class DataSeeder {
 
     private static Material material(String code, String name, String cas,
                                      Material.Category category, Material.Unit unit,
-                                     int unitPrice, int minOrder, int capacity, int leadTime,
+                                     long unitPrice, int minOrder, int capacity, int leadTime,
                                      List<String> certifications, String country,
                                      Long supplierId, String supplierName, String description) {
         return Material.builder()
@@ -109,7 +110,7 @@ public class DataSeeder {
                 .casNumber(cas)
                 .category(category)
                 .unit(unit)
-                .unitPrice(unitPrice)
+                .unitPrice(BigDecimal.valueOf(unitPrice))
                 .minOrderQuantity(minOrder)
                 .availableCapacity(capacity)
                 .leadTimeDays(leadTime)
